refactor: move utility actions to separate folder with test migration to localhost
