+++ src/actions/utility/index.ts
@@ -1,3 +1,4 @@
+export * from './closeVault';
 export * from './createExternalPriceAccount';
 export * from './createVault';
 export * from './initAuction';
