+++ test/actions/utility/initAuction.test.ts
@@ -2,6 +2,7 @@ import BN from 'bn.js';
 import { Keypair } from '@solana/web3.js';
 import { NATIVE_MINT } from '@solana/spl-token';
 import {
+  Auction,
   PriceFloor,
   PriceFloorType,
   WinnerLimit,
@@ -9,11 +10,12 @@ import {
 } from '@metaplex-foundation/mpl-auction';
 import { airdrop, LOCALHOST } from '@metaplex-foundation/amman';
 
+import { pause } from '../../utils';
 import { Connection, NodeWallet } from '../../../src';
 import { createExternalPriceAccount, createVault, initAuction } from '../../../src/actions/utility';
 
 describe('initAuction action', () => {
-  test('making an auction with newly created vault', async () => {
+  test('making an auction for newly created vault', async () => {
     const payer = Keypair.generate();
     const wallet = new NodeWallet(payer);
     const connection = new Connection(LOCALHOST, 'confirmed');
@@ -48,6 +50,11 @@ describe('initAuction action', () => {
       auctionSettings,
     });
 
-    expect(Boolean(auction)).not.toBeFalsy();
+    await pause(1000);
+
+    const auctionInstance = await Auction.load(connection, auction);
+
+    expect(auctionInstance).toHaveProperty('data');
+    expect(auctionInstance.data.tokenMint).toEqual(NATIVE_MINT.toBase58());
   });
 });
